requirements.txt
bottle==0.10.9
gunicorn==0.13.4
pymongo==2.1.1
Markdown==2.1.1
pbkdf2==1.3
requests==0.10.1
beautifulsoup4==4.0.0b3
Pygments==1.5
PyYAML==3.10
Jinja2==2.6